Major Update #2
~Energy System
+Time API
+Energy Update API
~ItemStackHelper

+++ src/main/java/io/github/mc_umod/api/energy/ICabel.java
@@ -13,5 +13,5 @@ public interface ICabel {
 		
 	public void setTunnel(int i);
 	
-	public double getRate();
+	public int getRate();
 }
